Add files via upload
You are working as a front end developer in an organization and the organization has job
openings in various departments. The organization has also received various applications for
these positions and the data is available in JSON format. Use the data shared along with this
problem statement. You, as a front end engineer, should develop a single User interface where
you can see the details shared by the candidates or the applicants.